requirements.txt
pandas
streamlit
python-dotenv
streamlit_option_menu 
streamlit_chat
africastalking
openai
seaborn
requests
matplotlib
numpy
folium
google-generativeai
streamlit-folium
folium
streamlit-stl
openrouteservices
plotly
geopy
sys
tempfile
langchain
langchain-community
langchain-core
langchain-google-genai
google-generativeai
python-dotenv
tqdm
chromadb
faiss-cpu
pypdf
